requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pyaml==20.4.0
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.8